+++ integration/esm-only-warning-test.ts
@@ -10,6 +10,7 @@ test.beforeAll(async () => {
 
   await createFixtureProject({
     buildStdio,
+    future: { v2_routeConvention: true },
     files: {
       "package.json": json({
         name: "remix-integration-9v4bpv66vd",
@@ -24,6 +25,7 @@ test.beforeAll(async () => {
           "@remix-run/node": "0.0.0-local-version",
           "@remix-run/react": "0.0.0-local-version",
           "@remix-run/serve": "0.0.0-local-version",
+          isbot: "0.0.0-local-version",
           react: "0.0.0-local-version",
           "react-dom": "0.0.0-local-version",
           "esm-only-no-exports": "0.0.0-local-version",
@@ -35,7 +37,7 @@ test.beforeAll(async () => {
           "@remix-run/dev": "0.0.0-local-version",
         },
       }),
-      "app/routes/index.jsx": js`
+      "app/routes/_index.jsx": js`
         import { json } from "@remix-run/node";
         import { Link, useLoaderData } from "@remix-run/react";
         import a from "esm-only-no-exports";