+++ integration/link-test.ts
@@ -32,6 +32,7 @@ test.describe("route module link export", () => {
 
   test.beforeAll(async () => {
     fixture = await createFixture({
+      future: { v2_routeConvention: true },
       files: {
         "app/favicon.ico": js``,
 
@@ -191,7 +192,7 @@ test.describe("route module link export", () => {
           }
         `,
 
-        "app/routes/index.jsx": js`
+        "app/routes/_index.jsx": js`
           import { useEffect } from "react";
           import { Link } from "@remix-run/react";
 
@@ -355,7 +356,7 @@ test.describe("route module link export", () => {
           }
         `,
 
-        "app/routes/gists/$username.jsx": js`
+        "app/routes/gists.$username.jsx": js`
           import { json, redirect } from "@remix-run/node";
           import { Link, useLoaderData, useParams } from "@remix-run/react";
           export async function loader({ params }) {
@@ -411,7 +412,7 @@ test.describe("route module link export", () => {
           }
         `,
 
-        "app/routes/gists/index.jsx": js`
+        "app/routes/gists._index.jsx": js`
           import { useLoaderData } from "@remix-run/react";
           export async function loader() {
             return ${JSON.stringify(fakeGists)};
@@ -452,7 +453,7 @@ test.describe("route module link export", () => {
           }
         `,
 
-        "app/routes/resources/theme-css.jsx": js`
+        "app/routes/resources.theme-css.jsx": js`
           import { redirect } from "@remix-run/node";
           export async function loader({ request }) {
             return new Response(":root { --nc-tx-1: #ffffff; --nc-tx-2: #eeeeee; }",
@@ -528,6 +529,35 @@ test.describe("route module link export", () => {
       await page.waitForSelector('[data-test-id="/responsive-image-preload"]');
       let locator = page.locator("link[rel=preload][as=image]");
       expect(await locator.getAttribute("imagesizes")).toBe("100vw");
+    });
+  });
+
+  test.describe("script imports", () => {
+    test("are added to the document", async ({ page }) => {
+      let app = new PlaywrightFixture(appFixture, page);
+      await app.goto("/");
+      let scripts = await page.$$("script");
+      expect(scripts.length).toEqual(2);
+      expect(await scripts[0].innerText()).toContain("__remixContext");
+      let moduleScript = scripts[1];
+      expect(await moduleScript.getAttribute("type")).toBe("module");
+      let moduleScriptText = await moduleScript.innerText();
+      expect(
+        Array.from(moduleScriptText.matchAll(/import "\/build\/manifest-/g)),
+        "invalid build manifest"
+      ).toHaveLength(1);
+      expect(
+        Array.from(moduleScriptText.matchAll(/import \* as route0 from "/g)),
+        "invalid route0"
+      ).toHaveLength(1);
+      expect(
+        Array.from(moduleScriptText.matchAll(/import \* as route1 from "/g)),
+        "invalid route1"
+      ).toHaveLength(1);
+      expect(
+        Array.from(moduleScriptText.matchAll(/import \* as route2 from "/g)),
+        "too many routes"
+      ).toHaveLength(0);
     });
   });
 });