+++ templates/deno/remix.config.js
@@ -1,7 +1,5 @@
 /** @type {import('@remix-run/dev').AppConfig} */
 module.exports = {
-  serverBuildTarget: "deno",
-  server: "./server.ts",
   /*
   If live reload causes page to re-render without changes (live reload is too fast),
   increase the dev server broadcast delay.
@@ -9,7 +7,16 @@ module.exports = {
   If live reload seems slow, try to decrease the dev server broadcast delay.
   */
   devServerBroadcastDelay: 300,
+  future: {
+    v2_routeConvention: true,
+  },
   ignoredRouteFiles: ["**/.*"],
+  server: "./server.ts",
+  serverConditions: ["deno", "worker"],
+  serverDependenciesToBundle: "all",
+  serverMainFields: ["module", "main"],
+  serverModuleFormat: "esm",
+  serverPlatform: "neutral",
   // appDirectory: "app",
   // assetsBuildDirectory: "public/build",
   // serverBuildPath: "build/index.js",