+++ packages/remix-cloudflare-workers/CHANGELOG.md
@@ -1,5 +1,19 @@
 # `@remix-run/cloudflare-workers`
 
+## 1.14.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/cloudflare@1.14.0`
+
+## 1.13.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/cloudflare@1.13.0`
+
 ## 1.12.0
 
 ### Patch Changes