+++ packages/remix-cloudflare/CHANGELOG.md
@@ -1,5 +1,19 @@
 # `@remix-run/cloudflare`
 
+## 1.14.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/server-runtime@1.14.0`
+
+## 1.13.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/server-runtime@1.13.0`
+
 ## 1.12.0
 
 ### Minor Changes