+++ packages/remix-express/CHANGELOG.md
@@ -1,5 +1,20 @@
 # `@remix-run/express`
 
+## 1.14.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/node@1.14.0`
+
+## 1.13.0
+
+### Patch Changes
+
+- Fix fetch `Request` creation for incoming URLs with double slashes ([#5336](https://github.com/remix-run/remix/pull/5336))
+- Updated dependencies:
+  - `@remix-run/node@1.13.0`
+
 ## 1.12.0
 
 ### Patch Changes